chore: Update debug and app URL configuration
The `debug` configuration in `config/app.php` has been updated to `false` to disable debugging in the application. Additionally, the `url` configuration in the same file has been updated to use `https://` as the default URL.

The `name` field in `package-lock.json` has been changed to "aplikasiprodi".

This commit message suggests that the changes made are related to updating the debug and app URL configuration, as well as changing the name field in `package-lock.json`.